[upd] разделение на проекты [upd] корректировка и обработка ошибок в ClientConnection::GetSocketAddress [add] красивый return в ProxyServer::StartServer с кодом ошибки, используя enum

+++ ProxyServer/IConnection.hpp
@@ -1,6 +1,7 @@
 #pragma once
 
-#include "Event.h"
+#include "Event.hpp"
+#include "ProxyServerApi.hpp"
 
 #define BUFF_SIZE 10240
 
@@ -10,8 +11,8 @@ public:
 	IConnection(SOCKET socket, sockaddr_in sock_addr, bool alive) : _socket(socket), _sock_addr(sock_addr), _alive(alive) {};
 	virtual ~IConnection() {};
 
-	virtual size_t GetData(char *buffer, size_t size_of_buffer, Event *_Event = nullptr) = 0;
-	virtual void SendData(const char *buffer, size_t len, Event *_Event = nullptr) = 0;
+	virtual int GetData(char *buffer, int size_of_buffer, Event *_Event = nullptr) = 0;
+	virtual void SendData(const char *buffer, int len, Event *_Event = nullptr) = 0;
 
 	inline bool isAlive() { return _alive; }
 	inline sockaddr_in GetSockaddr() { return _sock_addr; }
